+++ src/react/helpers/containerQuery.ts
@@ -1,0 +1,7 @@
+import {
+  getWidthClassesFromState as getWidthClasses,
+  observe,
+  unobserve,
+} from '../components/helpers/containerQuery';
+
+export { getWidthClasses, observe, unobserve };